Add files via upload
In the dataset, there has been a revenue decline for the bank, and they would like to know what action to take. So, the root cause is that their clients are not depositing as frequently as before. so, we have to predict whether a customer will subscribe to a term deposit or not.